+++ src/js/modules/utils/array-utils.js
@@ -1,4 +1,26 @@
+import {getProp} from "./object-utils";
+
 // noinspection JSUnusedGlobalSymbols
+
+/**
+ * Get an object from an array of objects based on a unique nested property value.
+ *
+ * @param arr {Array}
+ * @param prop {String} (Property in dot notation)
+ * @param value {*}
+ * @return {Object}
+ */
+export function getObjectByProperty(arr, prop, value) {
+
+    const obj = arr.find(item => getProp(item, prop) === value);
+
+    if (typeof obj !== "object") {
+        return {};
+    }
+
+    return obj;
+
+}
 
 /**
  * Sort an array of objects by a nested property value.