+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for `data-skin-youtube-id` and related attributes
 - Added `slug` method in `string-utils`
 - Added `aside` component
+- Added `getObjectByProperty` method in `array-utils`
 
 ### Changed